+++ include_sycl/int2e.hpp
@@ -256,7 +256,7 @@ inline int factorial2_gpu(int n) {
 }
 
 inline 
-long long int calcIdx4Dim(int i, int j, int k, int l, int nao)
+long long int calcIdx4Dim(size_t i, size_t j, size_t k, size_t l, size_t nao)
 {
     return nao * nao * nao * i + nao * nao * j + nao * k + l;
 }
